+++ internal/runtime/docker/container_network.go
@@ -3,7 +3,6 @@ package docker
 import (
 	"context"
 	"errors"
-	"fmt"
 
 	internalType "github.com/ErickMaria/envcontainer/internal/pkg/types"
 	"github.com/docker/docker/api/types"
@@ -11,10 +10,39 @@ import (
 	"github.com/docker/docker/api/types/network"
 )
 
-func (docker *Docker) createNetwork(ctx context.Context, options []internalType.Network) ([]string, error) {
+func (docker *Docker) createNetwork(ctx context.Context, options []internalType.Network, labels map[string]string) ([]string, error) {
+
+	getNetwork, err := docker.getNetwork(ctx, labels)
+	if err != nil {
+		return []string{}, err
+	}
+
+	// if len(getNetwork) != 0 {
+	// 	networkIds := []string{}
+
+	// 	// for _, net := range getNetwork {
+	// 	// 	networkIds = append(networkIds, net.ID)
+	// 	// }
+	// }
 
 	networkIds := []string{}
+	networkIdSkip := false
 	for _, netOpts := range options {
+
+		if len(getNetwork) != 0 {
+			for _, net := range getNetwork {
+
+				if netOpts.Name == net.Name {
+					networkIds = append(networkIds, net.ID)
+					networkIdSkip = true
+				}
+			}
+		}
+
+		if networkIdSkip {
+			networkIdSkip = false
+			continue
+		}
 
 		if netOpts.External {
 
@@ -32,9 +60,9 @@ func (docker *Docker) createNetwork(ctx context.Context, options []internalType.
 			}
 
 			if len(netList) == 0 {
-								return []string{}, errors.New("network with name "+netOpts.Name+" does not exist")
+				return []string{}, errors.New("network with name " + netOpts.Name + " does not exist")
 			}
-			
+
 			networkIds = append(networkIds, netList[0].ID)
 			continue
 		}
@@ -54,6 +82,7 @@ func (docker *Docker) createNetwork(ctx context.Context, options []internalType.
 			CheckDuplicate: true,
 			Driver:         netOpts.Driver,
 			IPAM:           &networkIPAMConfig,
+			Labels:         labels,
 		}
 
 		resp, err := docker.client.NetworkCreate(ctx, netOpts.Name, networkConfig)
@@ -63,7 +92,7 @@ func (docker *Docker) createNetwork(ctx context.Context, options []internalType.
 
 		networkIds = append(networkIds, resp.ID)
 
-		fmt.Printf("Created container %s on network %s: %s\n", resp.ID, netOpts.Name, resp.Warning)
+		// fmt.Printf("Created container %s on network %s: %s\n", resp.ID, netOpts.Name, resp.Warning)
 	}
 
 	return networkIds, nil